+++ test/unit/test_metas_store.py
@@ -3,13 +3,13 @@
 
 import uuid
 
-from voicevox_engine.metas.Metas import (
+from voicevox_engine.metas.metas import (
     SpeakerStyle,
     SpeakerSupportedFeatures,
     StyleId,
     StyleType,
 )
-from voicevox_engine.metas.MetasStore import (
+from voicevox_engine.metas.metas_store import (
     _SING_STYLE_TYPES,
     _TALK_STYLE_TYPES,
     Character,
