+++ test/unit/test_mock_tts_engine.py
@@ -1,7 +1,7 @@
 """MockTTSEngine クラスの単体テスト。"""
 
 from voicevox_engine.dev.tts_engine.mock import MockTTSEngine
-from voicevox_engine.metas.Metas import StyleId
+from voicevox_engine.metas.metas import StyleId
 from voicevox_engine.model import AudioQuery
 from voicevox_engine.tts_pipeline.kana_converter import create_kana
 from voicevox_engine.tts_pipeline.model import AccentPhrase, Mora
