+++ voicevox_engine/app/routers/character.py
@@ -6,8 +6,8 @@ from fastapi import APIRouter, Depends, HTTPException, Request
 from fastapi.responses import FileResponse
 from pydantic.json_schema import SkipJsonSchema
 
-from voicevox_engine.metas.Metas import Speaker, SpeakerInfo
-from voicevox_engine.metas.MetasStore import (
+from voicevox_engine.metas.metas import Speaker, SpeakerInfo
+from voicevox_engine.metas.metas_store import (
     Character,
     CharacterInfoNotFoundError,
     CharacterNotFoundError,
